+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1-alpha] - 2024-08-14
+
+### Added
+
+- Minimal implementation of `Shutdown()`
+  for `SocksClient` and `Command` (in case it handles communication).
+
 ## [0.3.0-alpha] - 2024-08-14
 
 ### Added
@@ -92,7 +99,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configurable SOCKS commands.
 - Configurable allow and block lists.
 
-[Unreleased]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.0-alpha...HEAD
+[Unreleased]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.1-alpha...HEAD
+[0.3.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.0-alpha...v0.3.1-alpha
 [0.3.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.2.0-alpha...v0.3.0-alpha
 [0.2.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.1.0-alpha...v0.2.0-alpha
 [0.1.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/releases/tag/v0.1.0-alpha